streamlit_app/requirements.txt
streamlit
opencv-python-headless==4.5.5.64
numpy
pillow
langdetect
deepl
pytesseract
dotenv
paddleocr
